@@ -13,13 +13,14 @@ namespace Tests\BitBag\SyliusCmsPlugin\Behat\Context\Api;
 use Behat\Behat\Context\Context;
 use Sylius\Behat\Client\ApiClientInterface;
 use Sylius\Behat\Client\ResponseCheckerInterface;
+use Tests\BitBag\SyliusCmsPlugin\Behat\Resources;
 use Webmozart\Assert\Assert;
 
 final class BlockContext implements Context
 {
-    private $apiClient;
+    private ApiClientInterface $apiClient;
 
-    private $responseChecker;
+    private ResponseCheckerInterface $responseChecker;
 
     public function __construct(
         ApiClientInterface $apiClient,
@@ -34,7 +35,7 @@ final class BlockContext implements Context
      */
     public function iWantToBrowseBlocks(): void
     {
-        $this->apiClient->index();
+        $this->apiClient->index(Resources::BLOCKS);
     }
 
     /**
@@ -58,7 +59,7 @@ final class BlockContext implements Context
     {
         Assert::true(
             $this->responseChecker->hasItemWithValue(
-                $this->apiClient->index(),
+                $this->apiClient->index(Resources::BLOCKS),
                 'code',
                 $code
             ),
