@@ -16,7 +16,7 @@ trait ResourceDataProcessingTrait
 {
     private function getResourceInterface(Request $request): object
     {
-        return null !== $request->get('id') && $this->repository->find($request->get('id')) ?
+        return null !== $request->get('id') && null !== $this->repository->find($request->get('id')) ?
             $this->repository->find($request->get('id')) :
             $this->factory->createNew();
     }
